+++ src/freenet/node/MessageWrapper.java
@@ -6,13 +6,13 @@ import freenet.support.SparseBitmap;
 public class MessageWrapper {
 	private final MessageItem item;
 	private final boolean isShortMessage;
-	private final int messageID;
+	private final long messageID;
 	
 	//Sorted lists of non-overlapping ranges
 	private final SparseBitmap acks = new SparseBitmap();
 	private final SparseBitmap sent = new SparseBitmap();
 
-	public MessageWrapper(MessageItem item, int messageID) {
+	public MessageWrapper(MessageItem item, long messageID) {
 		this.item = item;
 		isShortMessage = item.buf.length <= 255;
 		this.messageID = messageID;
@@ -58,7 +58,7 @@ public class MessageWrapper {
 		}
 	}
 
-	public int getMessageID() {
+	public long getMessageID() {
 		return messageID;
 	}
